feat: add windows portproxy documentation
without stopping the service first, ip helper will grab the port and
'listen' to it first before WSL gets the chance. the hacky workaround
for this is to just kill the service, bring up the container, then start
the service again. adding and removing the rule is probably unnecessary
but I figured it would help to do it anyways for safety.